refactor: move ContentContainer to ui
