+++ src/components/layouts/index.ts
@@ -5,13 +5,6 @@ import PolicyPageLayout from "./PolicyPageLayout";
 import type { PolicyPageLayoutProps } from "./PolicyPageLayout";
 import StickyScrollLayout from "./StickyScrollLayout";
 import type { StickyScrollLayoutProps } from "./StickyScrollLayout";
-import ContentContainer from "./ContentContainer";
 
-export {
-  PageBase,
-  PageHead,
-  PolicyPageLayout,
-  StickyScrollLayout,
-  ContentContainer,
-};
+export { PageBase, PageHead, PolicyPageLayout, StickyScrollLayout };
 export type { PageHeadProps, PolicyPageLayoutProps, StickyScrollLayoutProps };
