+++ src/pages/about.tsx
@@ -1,7 +1,7 @@
 import { FC, ReactElement, useEffect, useRef, useState } from "react";
 import dynamic from "next/dynamic";
 
-import { ContentContainer, PageBase, PageHead } from "@/components/layouts";
+import { PageBase, PageHead } from "@/components/layouts";
 import { useOnScreen } from "../hooks/useOnScreen";
 import { GetStaticProps } from "next";
 import { IClickUpTask } from "../types/clickUp";
@@ -11,6 +11,7 @@ import {
 } from "../lib/clickUp";
 import { TMemberDetails } from "../types/instill";
 import { OurMembersSection } from "@/components/sections";
+import { ContentContainer } from "@/components/ui";
 
 const SecureYourSpotSection = dynamic(() =>
   import("@/components/sections").then((mod) => mod.SecureYourSpotSection)
