+++ src/pages/career/[slug].tsx
@@ -3,10 +3,11 @@ import { useRouter } from "next/router";
 import { FC, ReactElement, useEffect, useRef, useState } from "react";
 import dynamic from "next/dynamic";
 
-import { ContentContainer, PageBase, PageHead } from "@/components/layouts";
+import { PageBase, PageHead } from "@/components/layouts";
+
 import { CareerPositionDetailsSection } from "@/components/career";
 
-import { BackToPreviousPageLink } from "@/components/ui";
+import { BackToPreviousPageLink, ContentContainer } from "@/components/ui";
 import {
   getClickUpTaskQuery,
   listClickUpTasksInListQuery,
