+++ src/pages/newsletter.tsx
@@ -4,7 +4,8 @@ import matter from "gray-matter";
 const mailchimp = require("@mailchimp/mailchimp_marketing");
 import { parse } from "node-html-parser";
 
-import { ContentContainer, PageBase, PageHead } from "@/components/layouts";
+import { PageBase, PageHead } from "@/components/layouts";
+import { ContentContainer } from "@/components/ui";
 import {
   GetCampaignContentResponse,
   ListCampaignsResponse,
